Setup Android SdlManager to use LifecycleManager
